Implemented uploading files with generating checsums for them and implemented CRUD methods for posts module

+++ src/users/users.controller.ts
@@ -7,10 +7,14 @@ import {
     Delete,
     UseInterceptors,
     ClassSerializerInterceptor,
+    Body,
+    NotFoundException,
+    BadRequestException,
 } from '@nestjs/common'
 import { JwtStrategyGuard } from 'src/auth/guards/jwt-auth.guard'
 import { UsersService } from './users.service'
 import { UserEntity } from './entities/user.entity'
+import { UpdateUserDto } from './dto/update-user.dto'
 
 @Controller('users')
 export class UsersController {
@@ -19,29 +23,26 @@ export class UsersController {
     @UseInterceptors(ClassSerializerInterceptor)
     @Get('profile')
     async getUser(@Request() req) {
-        const user = await this.usersService.getUser(req.user.username)
-        return {
-            id: user.id,
-            ...new UserEntity({
-                username: user.username,
-                firstname: user.firstname,
-                lastname: user.lastname,
-                email: user.email,
-                birthdate: user.birthdate,
-                phone: user.phone,
-            }),
-        }
+        const user = (
+            await this.usersService.getUser(req.user.username)
+        ).toObject()
+        return new UserEntity(user)
     }
     @UseGuards(JwtStrategyGuard)
     @UseInterceptors(ClassSerializerInterceptor)
     @Patch('profile/update')
-    async updateUser(@Request() req) {
-        await this.usersService.updateUser(req.user.userId, req.body.userData)
+    async updateUser(@Request() req, @Body() body: UpdateUserDto) {
+        if (!req.user.userId || !body) throw new BadRequestException()
+        return new UserEntity(
+            await this.usersService.updateUser(req.user.userId, body),
+        )
     }
     @UseGuards(JwtStrategyGuard)
     @Delete('profile/delete')
-    deleteUser(@Request() req) {
-        return this.usersService.deleteUser(req.user.userId)
+    async deleteUser(@Request() req) {
+        if (req.user.userId) throw new BadRequestException()
+        await this.usersService.deleteUser(req.user.userId)
+        return { message: 'User has been deleted' }
     }
 }
